Update sincronizacion de niebla
Falta cargar los tiles explorados en el cliente

+++ Diablo/Servidor/playerman.cpp
@@ -51,7 +51,7 @@ void Player::setOnline() {
 	isOnline = true;
 }
 
-void Player::addTileRecorrido(int x, int y) {
+void Player::addTileRecorrido(short x, short y) {
 	// ToDo: Esto se puede mejorar usando sort+binary search
 	auto p = std::make_pair(x, y);
 	if(std::find(tiles_recorridos.begin(), tiles_recorridos.end(), p) == tiles_recorridos.end()) {
